slack_size 적용에 필요한 모듈화 코드 생성하여 적용
