+++ python_engine/core/recovery/avi/extract_slack.py
@@ -6,6 +6,7 @@ from python_engine.core.recovery.avi.avi_split_channel import (
     extract_full_channel_bytes,
     CHUNK_SIG)
 from python_engine.core.recovery.utils.ffmpeg_wrapper import convert_video
+from python_engine.core.recovery.utils.unit import bytes_to_unit
 
 logger = logging.getLogger(__name__)
 
@@ -93,19 +94,6 @@ def extract_frames_from_raw(raw, sps_pps, out_fn):
             pos = nal_start
 
     return count, recovered_bytes
-
-def bytes_to_unit(n):
-    if n < 1024:
-        return f"{n} B"
-    elif n < 1024**2:
-        val = n / 1024
-        return f"{val:.1f} KB".rstrip("0").rstrip(".")
-    elif n < 1024**3:
-        val = n / 1024**2
-        return f"{val:.1f} MB".rstrip("0").rstrip(".")
-    else:
-        val = n / 1024**3
-        return f"{val:.1f} GB".rstrip("0").rstrip(".")
 
 def recover_avi_slack(input_avi, base_dir, target_format='mp4'):
     os.makedirs(base_dir, exist_ok=True)
